# backend/src/main/java/me/pacphi/ai/resos/controller/BookingController.java
package me.pacphi.ai.resos.controller;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import jakarta.validation.Valid;
import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;
import me.pacphi.ai.resos.model.Booking;
import me.pacphi.ai.resos.model.BookingsIdCommentPostRequest;
import me.pacphi.ai.resos.model.Error;
import me.pacphi.ai.resos.model.ValidationError;
import me.pacphi.ai.resos.repository.BookingRepository;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.time.LocalDate;
import java.util.List;

@RestController
@RequestMapping("/api/v1/resos")
public class BookingController {

    private final BookingRepository bookingRepository;

    public BookingController(BookingRepository bookingRepository) {
        this.bookingRepository = bookingRepository;
    }

    /**
     * GET /bookings/available-dates : Get available booking dates
     * Retrieve dates available for booking
     *
     * @return List of available dates (status code 200)
     */
    @Operation(
        operationId = "bookingsAvailableDatesGet",
        summary = "Get available booking dates",
        description = "Retrieve dates available for booking",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of available dates", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LocalDate.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @GetMapping(
        value = "/bookings/available-dates",
        produces = { "application/json" }
    )
    public ResponseEntity<List<LocalDate>> bookingsAvailableDatesGet() {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * GET /bookings/available-times : Get available booking times
     * Retrieve available times for a specific date
     *
     * @param date  (required)
     * @param people  (required)
     * @return List of available time slots (status code 200)
     */
    @Operation(
        operationId = "bookingsAvailableTimesGet",
        summary = "Get available booking times",
        description = "Retrieve available times for a specific date",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of available time slots", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = String.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @GetMapping(
        value = "/bookings/available-times",
        produces = { "application/json" }
    )
    public ResponseEntity<List<String>> bookingsAvailableTimesGet(
        @NotNull
        @Parameter(name = "date", description = "", required = true, in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "date", required = true)
        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date,

        @NotNull
        @Min(1)
        @Parameter(name = "people", description = "", required = true, in = ParameterIn.QUERY)
        @Valid
        @RequestParam(value = "people", required = true) Integer people
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * GET /bookings : List bookings
     * Retrieve a list of bookings with optional filtering
     *
     * @param fromDateTime ISO date format (ISO 8601) in UTC (optional)
     * @param toDateTime ISO date format (ISO 8601) in UTC (optional)
     * @param limit Number of records to return (max 100) (optional, default to 100)
     * @param skip Number of records to skip (optional, default to 0)
     * @param sort Sort field and direction (1 ascending, -1 descending) (optional)
     * @param customQuery Search expression for filtering results (optional)
     * @param onlyConfirmed Return only confirmed bookings (optional)
     * @return List of bookings (status code 200)
     *         or Unauthorized (status code 401)
     *         or Too many requests (status code 429)
     */
    @Operation(
        operationId = "bookingsGet",
        summary = "List bookings",
        description = "Retrieve a list of bookings with optional filtering",
        responses = {
            @ApiResponse(responseCode = "200", description = "List of bookings", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Booking.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "429", description = "Too many requests")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @GetMapping(
        value = "/bookings",
        produces = { "application/json" }
    )
    public ResponseEntity<List<Booking>> bookingsGet(
        @Parameter(name = "fromDateTime", description = "ISO date format (ISO 8601) in UTC", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "fromDateTime", required = false)
        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate fromDateTime,

        @Parameter(name = "toDateTime", description = "ISO date format (ISO 8601) in UTC", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "toDateTime", required = false)
        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate toDateTime,

        @Max(100)
        @Parameter(name = "limit", description = "Number of records to return (max 100)", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "limit", required = false, defaultValue = "100") Integer limit,

        @Parameter(name = "skip", description = "Number of records to skip", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "skip", required = false, defaultValue = "0") Integer skip,

        @Parameter(name = "sort", description = "Sort field and direction (1 ascending, -1 descending)", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "sort", required = false) String sort,

        @Parameter(name = "customQuery", description = "Search expression for filtering results", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "customQuery", required = false) String customQuery,

        @Parameter(name = "onlyConfirmed", description = "Return only confirmed bookings", in = ParameterIn.QUERY)
        @Valid @RequestParam(value = "onlyConfirmed", required = false) Boolean onlyConfirmed
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * POST /bookings/{id}/comment : Add comment to booking
     * Add a customer-visible comment to an existing booking
     *
     * @param id  (required)
     * @param bookingsIdCommentPostRequest  (required)
     * @return Comment added successfully (status code 200)
     *         or Booking not found (status code 404)
     */
    @Operation(
        operationId = "bookingsIdCommentPost",
        summary = "Add comment to booking",
        description = "Add a customer-visible comment to an existing booking",
        responses = {
            @ApiResponse(responseCode = "200", description = "Comment added successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "404", description = "Booking not found")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @PostMapping(
        value = "/bookings/{id}/comment",
        produces = { "application/json" },
        consumes = "application/json"
    )
    public ResponseEntity<String> bookingsIdCommentPost(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH)
        @PathVariable("id") String id,

        @Parameter(name = "BookingsIdCommentPostRequest", description = "", required = true)
        @Valid @RequestBody BookingsIdCommentPostRequest bookingsIdCommentPostRequest
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * GET /bookings/{id} : Get booking
     * Retrieve a specific booking by ID
     *
     * @param id  (required)
     * @return Booking details (status code 200)
     *         or Booking not found (status code 404)
     */
    @Operation(
        operationId = "bookingsIdGet",
        summary = "Get booking",
        description = "Retrieve a specific booking by ID",
        responses = {
            @ApiResponse(responseCode = "200", description = "Booking details", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Booking.class))
            }),
            @ApiResponse(responseCode = "404", description = "Booking not found")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @GetMapping(
        value = "/bookings/{id}",
        produces = { "application/json" }
    )
    public ResponseEntity<Booking> bookingsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH)
        @PathVariable("id") String id
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * PUT /bookings/{id} : Update booking
     * Update an existing booking
     *
     * @param id  (required)
     * @param booking  (required)
     * @return Booking updated successfully (status code 200)
     *         or Booking not found (status code 404)
     */
    @Operation(
        operationId = "bookingsIdPut",
        summary = "Update booking",
        description = "Update an existing booking",
        responses = {
            @ApiResponse(responseCode = "200", description = "Booking updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Boolean.class))
            }),
            @ApiResponse(responseCode = "404", description = "Booking not found")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @PutMapping(
        value = "/bookings/{id}",
        produces = { "application/json" },
        consumes = "application/json"
    )
    public ResponseEntity<Boolean> bookingsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH)
        @PathVariable("id") String id,

        @Parameter(name = "Booking", description = "", required = true)
        @Valid @RequestBody Booking booking
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * POST /bookings/{id}/restaurantNote : Add note to booking
     * Add an internal note to an existing booking
     *
     * @param id  (required)
     * @param bookingsIdCommentPostRequest  (required)
     * @return Note added successfully (status code 200)
     *         or Booking not found (status code 404)
     */
    @Operation(
        operationId = "bookingsIdRestaurantNotePost",
        summary = "Add note to booking",
        description = "Add an internal note to an existing booking",
        responses = {
            @ApiResponse(responseCode = "200", description = "Note added successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = String.class))
            }),
            @ApiResponse(responseCode = "404", description = "Booking not found")
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @PostMapping(
            value = "/bookings/{id}/restaurantNote",
            produces = { "application/json" },
            consumes = "application/json"
    )
    public ResponseEntity<String> bookingsIdRestaurantNotePost(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH)
        @PathVariable("id") String id,

        @Parameter(name = "BookingsIdCommentPostRequest", description = "", required = true)
        @Valid @RequestBody BookingsIdCommentPostRequest bookingsIdCommentPostRequest
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }

    /**
     * POST /bookings : Create booking
     * Create a new booking
     *
     * @param booking  (required)
     * @return Booking created successfully (status code 200)
     *         or Bad request - invalid booking data (status code 400)
     *         or Unauthorized - invalid or missing API key (status code 401)
     *         or Unprocessable Entity - e.g., table not available (status code 422)
     *         or Too many requests - rate limit exceeded (status code 429)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "bookingsPost",
        summary = "Create booking",
        description = "Create a new booking",
        responses = {
            @ApiResponse(responseCode = "200", description = "Booking created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Booking.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request - invalid booking data", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ValidationError.class))
            }),
            @ApiResponse(responseCode = "401", description = "Unauthorized - invalid or missing API key", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = me.pacphi.ai.resos.model.Error.class))
            }),
            @ApiResponse(responseCode = "422", description = "Unprocessable Entity - e.g., table not available", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = me.pacphi.ai.resos.model.Error.class))
            }),
            @ApiResponse(responseCode = "429", description = "Too many requests - rate limit exceeded", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = me.pacphi.ai.resos.model.Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "basicAuth")
        }
    )
    @PostMapping(
        value = "/bookings",
        produces = { "application/json" },
        consumes = "application/json"
    )
    public ResponseEntity<Booking> bookingsPost(
        @Parameter(name = "Booking", description = "", required = true)
        @Valid @RequestBody Booking booking
    ) {
        throw new UnsupportedOperationException("Not implemented");
    }
}