# client/src/main/resources/openapi/resos-openapi.yml
openapi: 3.0.0
info:
  title: ResOS API
  version: '1.2'
  description: |
    ResOS API for managing restaurant bookings, tables, opening hours, orders, feedback and customers.
    
    For support or feature requests, contact: hi@resos.com
  contact:
    email: hi@resos.com
servers:
  - url: https://api.resos.com/v1
    description: Production API Server
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Authentication using API key as username (no password required).
        The API key must be base64 encoded with a trailing colon.
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - AUTHENTICATION_ERROR
            - AUTHORIZATION_ERROR
            - RESOURCE_NOT_FOUND
            - VALIDATION_ERROR
            - RATE_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - CONFLICT
            - SERVICE_UNAVAILABLE
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details specific to the error type
    ValidationError:
      type: object
      required:
        - code
        - message
        - fields
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR]
        message:
          type: string
        fields:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string
...
    Guest:
      type: object
      required:
        - name
      description: |
        Represents a guest making a booking. Contains all necessary contact information
        and notification preferences for the guest.
      properties:
        name:
          type: string
          example: "Thomas A. Anderson"
          minLength: 2
          maxLength: 100
          description: "Full name of the guest making the booking"
        email:
          type: string
          format: email
          example: "neo@resos.com"
          description: "Valid email address for booking confirmations and updates"
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$'
        phone:
          type: string
          pattern: '^\\+[1-9][0-9]{6,14}$'
          example: "+13115552368"
          description: |
            Phone number in E.164 format, including country code with + prefix.
            Must contain only numbers after the + symbol.
        notificationSms:
          type: boolean
          default: false
          description: |
            When true, guest will receive booking confirmations and updates via SMS.
            Requires a valid phone number to be provided.
        notificationEmail:
          type: boolean
          default: false
          description: |
            When true, guest will receive booking confirmations and updates via email.
            Requires a valid email address to be provided.
    Address:
      type: object
      properties:
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        postalCode:
          type: integer
        countryCode:
          type: string
          example: "DK"
        country:
          type: string
          example: "Denmark"
    Restaurant:
      type: object
      properties:
        name:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        email:
          type: string
          format: email
        phone:
          type: string
        slug:
          type: string
        languageCode:
          type: string
          enum: [en, es, de, da, sv]
          default: en
    Area:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        internalNote:
          type: string
        bookingPriority:
          type: integer
    Table:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        seatsMin:
          type: integer
          minimum: 1
        seatsMax:
          type: integer
        area:
          $ref: '#/components/schemas/Area'
        internalNote:
          type: string
    Booking:
      type: object
      required:
        - date
        - time
        - people
        - duration
        - guest
        - openingHourId
      description: |
        Represents a restaurant booking. Contains all necessary information about
        the reservation, including guest details, timing, and table allocation.
      properties:
        _id:
          type: string
          description: "Unique identifier for the booking"
          readOnly: true
        date:
          type: string
          format: date
          example: "2021-09-18"
          description: "Date of the booking in YYYY-MM-DD format"
        time:
          type: string
          pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "21:00"
          description: "Time of the booking in 24-hour HH:mm format"
        people:
          type: integer
          minimum: 1
          maximum: 999
          description: "Number of people in the booking party"
        duration:
          type: integer
          minimum: 15
          maximum: 480
          description: |
            Duration of the booking in minutes.
            Minimum 15 minutes, maximum 8 hours (480 minutes).
        status:
          type: string
          enum:
            - request
            - declined
            - approved
            - arrived
            - seated
            - left
            - no_show
            - canceled
            - waitlist
          default: request
          description: |
            Current status of the booking:
            * request - Initial booking request
            * declined - Booking was declined by the restaurant
            * approved - Booking was approved by the restaurant
            * arrived - Guest has arrived at the restaurant
            * seated - Guest has been seated at their table
            * left - Guest has finished their meal and left
            * no_show - Guest did not show up for their booking
            * canceled - Booking was canceled by the guest
            * waitlist - Booking is on the waitlist
        tables:
          type: array
          items:
            $ref: '#/components/schemas/Table'
          description: "Array of tables allocated for this booking"
        metadata:
          type: object
          description: |
            Contains any arbitrary metadata for the booking.
            Maximum total length of stringified JSON is 2000 characters.
          maxLength: 2000
        comments:
          type: array
          items:
            type: string
          description: "Array of customer-visible comments on the booking"
        guest:
          $ref: '#/components/schemas/Guest'
        source:
          type: string
          enum:
            - walkin
            - website
            - phone
            - email
            - message
            - eatie
            - other
          default: other
          description: |
            Source of the booking:
            * walkin - Guest walked in without a booking
            * website - Booked through the restaurant's website
            * phone - Booked via phone call
            * email - Booked via email
            * message - Booked via messaging service
            * eatie - Booked through Eatie platform
            * other - Other booking source
        openingHourId:
          type: string
          description: "Reference to the opening hours period this booking is for"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: "Timestamp when the booking was created"
        activityAt:
          type: string
          format: date-time
          readOnly: true
          description: "Timestamp of the last activity on this booking"
        sendNotification:
          type: boolean
          description: "Whether notifications should be sent for this booking"
        payment:
          type: object
          description: |
            Payment information for bookings with prepayment or no-show fee.
            Structure depends on the payment provider used.
        comment:
          type: string
          maxLength: 1000
          description: |
            Public comment written by guest.
            Visible to both the guest and the restaurant.
        note:
          type: string
          maxLength: 1000
          description: "Internal note only visible to the restaurant"
        noteAuthor:
          type: string
          default: "API"
          maxLength: 100
          description: "Name of the person who wrote the internal note"
        referrer:
          type: string
          format: uri
          maxLength: 2000
          description: "URL of website that referred the booking"
        languageCode:
          type: string
          enum: [en, es, de, da, sv]
          default: en
          description: |
            Language code for guest communications:
            * en - English
            * es - Spanish
            * de - German
            * da - Danish
            * sv - Swedish
        restaurant:
          $ref: '#/components/schemas/Restaurant'
  parameters:
    fromDateTime:
      name: fromDateTime
      in: query
      description: ISO date format (ISO 8601) in UTC
      required: false
      schema:
        type: string
        format: date
        example: "2021-04-17"
    toDateTime:
      name: toDateTime
      in: query
      description: ISO date format (ISO 8601) in UTC
      required: false
      schema:
        type: string
        format: date
        example: "2021-04-18"
    limit:
      name: limit
      in: query
      description: Number of records to return (max 100)
      required: false
      schema:
        type: integer
        default: 100
        maximum: 100
    skip:
      name: skip
      in: query
      description: Number of records to skip
      required: false
      schema:
        type: integer
        default: 0
    sort:
      name: sort
      in: query
      description: Sort field and direction (1 ascending, -1 descending)
      required: false
      schema:
        type: string
        example: "createdAt:1"
    customQuery:
      name: customQuery
      in: query
      description: Search expression for filtering results
      required: false
      schema:
        type: string
        example: "people>1,duration>60"
paths:
...
  /bookings:
    get:
      summary: List bookings
      description: Retrieve a list of bookings with optional filtering
      parameters:
        - $ref: '#/components/parameters/fromDateTime'
        - $ref: '#/components/parameters/toDateTime'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/skip'
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/customQuery'
        - name: onlyConfirmed
          in: query
          description: Return only confirmed bookings
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized
        '429':
          description: Too many requests
          headers:
            Retry-After:
              description: Time in ms until reset
              schema:
                type: integer

    post:
      summary: Create booking
      description: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '200':
          description: Booking created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Bad request - invalid booking data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized - invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Unprocessable Entity - e.g., table not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests - rate limit exceeded
          headers:
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bookings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: Get booking
      description: Retrieve a specific booking by ID
      responses:
        '200':
          description: Booking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '404':
          description: Booking not found

    put:
      summary: Update booking
      description: Update an existing booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Booking'
      responses:
        '200':
          description: Booking updated successfully
          content:
            application/json:
              schema:
                type: boolean
        '404':
          description: Booking not found
...
  /bookings/available-times:
    get:
      summary: Get available booking times
      description: Retrieve available times for a specific date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: people
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: List of available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  pattern: '^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$'
  /bookings/{id}/comment:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Add comment to booking
      description: Add a customer-visible comment to an existing booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
      responses:
        '200':
          description: Comment added successfully
          content:
            application/json:
              schema:
                type: string
                description: Comment ID
        '404':
          description: Booking not found
  /bookings/{id}/restaurantNote:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    post:
      summary: Add note to booking
      description: Add an internal note to an existing booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
      responses:
        '200':
          description: Note added successfully
          content:
            application/json:
              schema:
                type: string
                description: Note ID
        '404':
          description: Booking not found